+++ app/lib/actions/debt.ts
@@ -2,14 +2,16 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import prisma from "@/app/lib/prisma";
-import { DebtFormErrorState, DebtSchema } from "../zod-schemas";
+import {
+    CreateDebtFormSchema,
+    DebtFormError,
+    DebtFormSchema,
+    IdSchema,
+} from "../zod-schemas";
 
-export async function createDebt(
-    prevState: DebtFormErrorState,
-    formData: FormData
-) {
+export async function createDebt(prevState: DebtFormError, formData: FormData) {
     // Validate form fields using Zod
-    const validatedFields = DebtSchema.safeParse({
+    const validatedFields = CreateDebtFormSchema.safeParse({
         name: formData.get("name"),
         amount: formData.get("amount"),
         interest: formData.get("interest"),
@@ -44,7 +46,7 @@ export async function createDebt(
             message: "Database Error: Failed to Create Debt.",
         };
     }
-    
+
     // Revalidate the cache
     revalidatePath("/dashboard/debts");
 
@@ -53,12 +55,13 @@ export async function createDebt(
 }
 
 export async function updateDebt(
-    id: string,
-    prevState: DebtFormErrorState,
+    debtId: string,
+    prevState: DebtFormError,
     formData: FormData
 ) {
     // Validate form fields using Zod
-    const validatedFields = DebtSchema.safeParse({
+    const validatedFields = DebtFormSchema.safeParse({
+        id: debtId,
         name: formData.get("name"),
         amount: formData.get("amount"),
         interest: formData.get("interest"),
@@ -74,7 +77,7 @@ export async function updateDebt(
     }
 
     // Extract validated fields
-    const { name, amount, interest, categoryId } = validatedFields.data;
+    const { id, name, amount, interest, categoryId } = validatedFields.data;
 
     // Update the debt
     try {
@@ -86,7 +89,7 @@ export async function updateDebt(
                 name: name,
                 amount: amount,
                 interest: interest,
-                categoryId: categoryId
+                categoryId: categoryId,
             },
         });
     } catch (error) {
@@ -104,13 +107,26 @@ export async function updateDebt(
 }
 
 export async function deleteDebt(id: string) {
+    // Validate the id
+    const parsedId = IdSchema.safeParse(id);
+
+    // If ID validation fails, throw an error
+    if (!parsedId.success) {
+        console.error("Invalid ID format:", parsedId.error);
+        throw new Error("Invalid ID format");
+    }
+
+    // Extract the validated ID
+    const validatedId = parsedId.data;
+
     try {
         await prisma.debt.delete({
             where: {
-                id,
+                id: validatedId,
             },
         });
 
+        // Revalidate the cache
         revalidatePath("/dashboard/debts");
     } catch (error) {
         console.error("Failed to delete debt:", error);
